chore: refactor tracingmiddelware
# What does this PR do?


## Test Plan
